+++ packages/react/src/common/contexts/MenuExpandContext/MenuExpandContext.definitions.ts
@@ -1,7 +1,10 @@
+import { noop } from "@common/utils";
 import type { MenuExpandContextValue } from "./MenuExpandContext.types";
 
 export const initialMenuContext: MenuExpandContextValue = {
   expanded: true,
-  showTooltipsWhenCollapsed: true,
-  tooltipsPosition: "right",
-} as const;
+  showPopoversWhenCollapsed: true,
+  popoverPosition: "right",
+  activeAccordionPopover: null,
+  setActiveAccordionPopover: noop,
+};