+++ packages/react/src/components/Menu/src/components/MenuBody/menuBody.spec.tsx
@@ -2,14 +2,24 @@ import React from "react";
 import { render, screen } from "@testing-library/react";
 
 import { MenuExpandContext } from "@common/contexts";
+import { noop } from "@common/utils";
 import { MenuBody } from "./MenuBody";
 import { MenuBodyProps } from "./menuBody.types";
 
 const bodyChildren = "Body content";
 
-const makeSut = (rest: Omit<MenuBodyProps, "children">) => {
+const makeSut = (
+  rest: Omit<MenuBodyProps, "children">,
+  contextExpanded = true
+) => {
   render(
-    <MenuExpandContext.Provider value={{ expanded: true }}>
+    <MenuExpandContext.Provider
+      value={{
+        expanded: contextExpanded,
+        activeAccordionPopover: null,
+        setActiveAccordionPopover: noop,
+      }}
+    >
       <MenuBody {...rest} data-testid="body-element">
         {bodyChildren}
       </MenuBody>
@@ -21,6 +31,11 @@ describe("GIVEN <MenuBody />", () => {
   describe("WHEN rendered", () => {
     it("SHOULD render children correctly", () => {
       makeSut({});
+      expect(screen.getByText(bodyChildren)).toBeDefined();
+    });
+
+    it("SHOULD render children when collapsed", () => {
+      makeSut({}, false);
       expect(screen.getByText(bodyChildren)).toBeDefined();
     });
   });