+++ packages/react/src/components/MenuButton/src/components/MenuButtonContent/index.ts
@@ -1,0 +1,2 @@
+export { MenuButtonContent } from "./MenuButtonContent";
+export type { MenuButtonContentProps } from "./menuButtonContent.types";