+++ packages/react/src/components/MenuButton/src/components/index.ts
@@ -1,2 +1,4 @@
 export * from "./MenuButtonAccordion";
 export type { MenuButtonAccordionBaseProps } from "./MenuButtonAccordion/menuButtonAccordion.types";
+export * from "./MenuButtonContent";
+export type { MenuButtonContentProps } from "./MenuButtonContent/menuButtonContent.types";